+++ wp-content/themes/Parallax-One/cashier_app/cashier_app_form.php
@@ -1,7 +1,22 @@
 <?php
-
 // =============================
 // CASHIER APP - V1.0 Form part
+
+// Process part: cashier_app_process.php (in the root folder)
+// Summary part: cashier_app_summary.php (in the theme folder)
+
+// OUTLINE
+// 1. FUNCTIONS
+// 2. INITIALIZATION
+// 3. FORM
+// =============================
+
+require_once(ABSPATH . 'wp-config.php');
+$connection_4w = mysqli_connect(DB_HOST, DB_USER, DB_PASSWORD);
+mysqli_select_db($connection_4w, DB_NAME);
+
+// =============================
+// 1. FUNCTIONS
 // =============================
 
 // @return closest_lesson [Array] Array containing: Timestamp of the closest lesson, Class type, Level
@@ -35,9 +50,9 @@ function get_closest_lesson($connection_4w, $branch_id) {
 	return $result_array;
 }
 
-	require_once(ABSPATH . 'wp-config.php');
-	$connection_4w = mysqli_connect(DB_HOST, DB_USER, DB_PASSWORD);
-	mysqli_select_db($connection_4w, DB_NAME);
+// =============================
+// 2. INITIALIZATION
+// =============================
 
 	// Find branch ID from URL
 	$url = parse_url( $_SERVER['REQUEST_URI'], PHP_URL_PATH );
@@ -77,11 +92,15 @@ function get_closest_lesson($connection_4w, $branch_id) {
 		$filled_class_type = $_POST['class'];
 		$filled_level = $_POST['level'];
 	}
+
+// =============================
+// 3. FORM
+// =============================
 ?>
 	<div class="cashier">
 		<br />
 		<link rel="stylesheet" href="../wp-content/themes/Parallax-One/cashier_app/cashier_app.css">
-		<form action="/index.php" id="cashier" method="post">
+		<form action="/cashier_app_process.php" id="cashier" method="post">
 			<p>Welcome, dear cashier. On this page you count the number of sold 1-time entries, vouchers, and how many people came on already-paid vouchers.</p>
 			<p>Please check the class, level, date and time.</p>
 			<div class="cashier-upper col-md-12">
